many improvement to generation!

+++ src/main/java/net/gyula/wildaside/procedures/EntoriumShearingProcedure.java
@@ -60,11 +60,11 @@ public class EntoriumShearingProcedure {
 				if (EnchantmentHelper.getItemEnchantmentLevel(WildasideModEnchantments.REVEALING.get(), (entity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY)) != 0) {
 					sx = -1;
 					found = false;
-					for (int index0 = 0; index0 < (int) (3); index0++) {
+					for (int index0 = 0; index0 < 3; index0++) {
 						sy = -1;
-						for (int index1 = 0; index1 < (int) (3); index1++) {
+						for (int index1 = 0; index1 < 3; index1++) {
 							sz = -1;
-							for (int index2 = 0; index2 < (int) (3); index2++) {
+							for (int index2 = 0; index2 < 3; index2++) {
 								if ((world.getBlockState(new BlockPos(x + sx, y + sy, z + sz))).getBlock() == WildasideModBlocks.OVERGROWN_ENTORIUM_ORE.get()) {
 									{
 										BlockPos _bp = new BlockPos(x + sx, y + sy, z + sz);
