+++ src/main/java/net/gyula/wildaside/procedures/GlowingEssenceLeavesTransformProcProcedure.java
@@ -84,11 +84,11 @@ public class GlowingEssenceLeavesTransformProcProcedure {
 				}
 				sx = -1;
 				found = false;
-				for (int index0 = 0; index0 < (int) (3); index0++) {
+				for (int index0 = 0; index0 < 3; index0++) {
 					sy = -1;
-					for (int index1 = 0; index1 < (int) (3); index1++) {
+					for (int index1 = 0; index1 < 3; index1++) {
 						sz = -1;
-						for (int index2 = 0; index2 < (int) (3); index2++) {
+						for (int index2 = 0; index2 < 3; index2++) {
 							if ((world.getBlockState(new BlockPos(x + sx, y + sy, z + sz))).getBlock() == WildasideModBlocks.HICKORY_LEAVES.get()) {
 								{
 									BlockPos _bp = new BlockPos(x + sx, y + sy, z + sz);
